pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "commitmessagegenerator"
version = "0.1.5"
description = "Generate commit messages with AI (Google Gemini) automatically using `git diff`."
readme = "README.md"
license = {text = "MIT"}
authors = [
  {name = "Gabriel Terceiro", email = "gcarolinoterceiro@gmail.com"}
]
dependencies = [
    "google-generativeai",
    "GitPython",
    "python-dotenv"
]

[project.scripts]
commitgen = "commitmessagegenerator.cli:main"
